Rename the unclear "desktop control" module to "controller"
This includes dropping the platform-specific "desktop" in favor of
the more platform-generic "display" for the DC backends as we could
technically manipulate Android or other devices.
